Increase react-testing async timeout
Also, decrease jest timeout. Set react-testing timeout to be 1s less
than jest.

We saw consistent test failures when running search results view
integration test on its own, or when it happened to run first in the
suite. The suspense fallback would still be visible by the timeout.

We think that this commit changed how fast this happened, but we don't
know why:

https://github.com/boclips/publishers/commit/d3f6e9a45f6c8bb7b2ebe847a45f9e0ebf0ac6f2

Signed-off-by: Matt Jones <matt@boclips.com>