Use debounce when applying / clearing filters
Apologies, could not figure out a way to test this behaviour.
Originally wanted to link useDebounce to the filters coming from the
url, but it caused infinite debounce calls with the same values. So
instead added `newFiltersBeforeDebounce` to the middle which we only
update when the `handleFilterChange` or `removeAllFilters`
callbacks are run and then debounce that value.

[#175576578]