Introduce a catch all error boundary
This catches all network/js errors and shows a fallback if anything
occurs.
Not in love with the network error boundary as I think it could cause
confusion. The way it currently works is listens to the query cache and
looks for any queries in the error state. It basically means you have to
refresh the page before going back to a working app. It also means we
don't have fine control currently of handling the error if we want to
display something else. If this is the case you'd need to not wrap the
all knowing error boundary so you can handle your error explicitly

[#176804377]